requirements.txt
# Core ML Framework
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
xgboost==1.7.6
lightgbm==4.0.0

# Deep Learning
torch==2.0.1
tensorflow==2.13.0

# Distributed Training & Orchestration
ray[default]==2.6.3
mlflow==2.5.0

# Data Processing
apache-airflow==2.6.3
apache-airflow-providers-apache-spark==4.1.3
pyspark==3.4.1
kafka-python==2.0.2
confluent-kafka==2.2.0

# Feature Store
feast==0.32.0
redis==4.6.0
psycopg2-binary==2.9.6

# Data Quality
great-expectations==0.17.8

# API & Serving
fastapi==0.100.0
uvicorn[standard]==0.23.1
pydantic==2.0.3
python-multipart==0.0.6

# Monitoring & Observability
prometheus-client==0.17.1
evidently==0.4.3
grafana-api==1.0.3

# Database & Caching
sqlalchemy==2.0.19
alembic==1.11.1
redis==4.6.0

# Cloud & Infrastructure
boto3==1.28.17
google-cloud-storage==2.10.0
kubernetes==27.2.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
pyyaml==6.0.1
click==8.1.6
jsonschema==4.18.4

# Testing
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.24.1

# Security & Compliance
cryptography==41.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development
black==23.7.0
flake8==6.0.0
mypy==1.4.1
pre-commit==3.3.3
